@@ -16,10 +16,11 @@
 package com.example.android.viewpager;
 
 import android.os.Bundle;
-import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+
+import androidx.fragment.app.Fragment;
 
 /**
  * Fragment that displays "Tuesday".
